# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comic Canvas API",
    "version": "0.1.0"
  },
  "paths": {
    "/media/upload": {
      "put": {
        "tags": [
          "media"
        ],
        "summary": "Upload",
        "operationId": "upload_media_upload_put",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Expires"
            }
          },
          {
            "name": "signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/media/{key}": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Serve",
        "operationId": "serve_media__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/weekly/{iso_week}/generate": {
      "post": {
        "tags": [
          "weekly"
        ],
        "summary": "Generate Weekly",
        "operationId": "generate_weekly_weekly__iso_week__generate_post",
        "parameters": [
          {
            "name": "iso_week",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iso Week"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Generate Weekly Weekly  Iso Week  Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/weekly/{iso_week}.pdf": {
      "get": {
        "tags": [
          "weekly"
        ],
        "summary": "Get Weekly Pdf",
        "operationId": "get_weekly_pdf_weekly__iso_week__pdf_get",
        "parameters": [
          {
            "name": "iso_week",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iso Week"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/weekly/{iso_week}": {
      "get": {
        "tags": [
          "weekly"
        ],
        "summary": "Get Weekly",
        "operationId": "get_weekly_weekly__iso_week__get",
        "parameters": [
          {
            "name": "iso_week",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iso Week"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Weekly Weekly  Iso Week  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}